--- src/OnBoardingPage.tsx
@@ -1,5 +1,0 @@
-const OnBoardingPage = () => {
-	return <div>OnBoardingPage</div>;
-};
-
-export default OnBoardingPage;
